backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
